requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib
python-multipart
cryptography
requests
python-dotenv